fixed test cases
