+++ src/components/chapter/Chapter.scss
@@ -67,79 +67,6 @@
   padding: 0 30px;
 }
 
-.right-panel {
-  width: 0;
-  background-color: #fff;
-  border-left: 1px solid #e0e0e0;
-  transition: width 0.3s ease-in-out;
-  overflow: auto;
-  box-shadow: -2px 0 5px rgba(0, 0, 0, 0.05);
-  height: 100%;
-  
-  &.show {
-    width: 400px;
-  }
-
-  .headerthing{
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    background-color: #f0f0f0;
-  }
-  .close-icon {
-    cursor: pointer;
-    padding: 4px;
-    border-radius: 50%;
-    background-color: #f0f0f0;
-    color: #333;
-    margin-right: 10px;
-    transition: all 0.2s ease-in-out;
-    
-    &:hover {
-      background-color: #e0e0e0;
-      transform: scale(1.1);
-    }
-  }
-
-  .panel-content {
-    text-align: start;
-    opacity: 0;
-    transition: opacity 0.3s ease-in-out;
-    
-    .textgreat{
-      color: #828282;
-      border-bottom: 1px solid #f0f0f0;
-    }
-    h3 {
-      margin-bottom: 20px;
-      color: #333;
-      font-weight: 600;
-    }
-    
-    p {
-      display: flex;
-      align-items: center;
-      padding: 12px 0;
-      margin: 0;
-      color: #444;
-      cursor: pointer;
-      transition: all 0.2s ease;
-      
-      &:not(.textgreat):hover {
-        color: #7b7b7b;
-        background-color: #f8f9fa;
-      }
-
-     
-    }
-
-    
-  }
-  
-  &.show .panel-content {
-    opacity: 1;
-  }
-}
 
 /* Style for citation text */
 .text-citation-style {
@@ -180,72 +107,8 @@
   .tibetan-text {
     font-size: 1rem;
   }
-  .right-panel.show {
-    width: 320px;
-  }
 }
 
-.share-content {
-  text-align: start;
-  .textgreat{
-    color: #828282;
-    border-bottom: 1px solid #f0f0f0;
-  }
-
-  .social-share-buttons {
-    margin-top: 16px;
-    
-    .social-button {
-      display: flex;
-      align-items: center;
-      padding: 12px 16px;
-      margin: 8px 0;
-      cursor: pointer;
-      transition: background-color 0.2s ease;
-      border-radius: 4px;
-
-      &:hover {
-        background-color: #f5f5f5;
-      }
-
-      .social-icon {
-        color: #666;
-        margin-right: 12px;
-        font-size: 18px;
-      }
-    }
-  }
-
-  .share-url-container {
-    display: flex;
-    align-items: center;
-    background-color: #f5f5f5;
-    border-radius: 4px;
-    gap: 10px;
-
-    .share-url {
-      flex: 1;
-      margin: 0;
-      font-size: 14px;
-      color: #666;
-      overflow: hidden;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-    }
-
-    .copy-button {
-      background: none;
-      border: none;
-      padding: 5px;
-      cursor: pointer;
-      color: #666;
-      
-      &:hover {
-        color: #333;
-      }
-    }
-  }
-}
 
 .segment {
     position: relative;
